Introduce t_infomask2 bits for partial heap only tuples.
This change introduces HEAP_PHOT_UPDATED and HEAP_PHOT_TUPLE, which
will be used to identify whether a tuple was PHOT-updated or is a
partial heap-only tuple itself.  Related macros are also introduced
to match the existing ones for HOT.  No code actually makes use of
these bits yet.

Note that this change lays claim to the last two available bits in
t_infomask2, which may add a considerable amount of complexity for
future heap changes (to retain data format compatibility).

Signed-off-by: Greg Burd <greg@burd.me>