Add logic for following basic partial heap only tuple chains.
The bulk of this change is for teaching heap_hot_search_buffer() how
to follow PHOT chains (with or without a bitmap of index attributes to
check against).  Since such index attribute bitmaps are more difficult
to discover and use for bitmap scans, this change also teaches the
bitmap scan logic how to handle PHOT chains.  Specifically, the
recheck condition is used liberally, extra checks are in place to
avoid duplicating tuple offsets in result arrays, and tbm_intersect()
is modified to OR page bitmaps together to account for index tuples
that point to different parts of the same PHOT chain.

Several notes are sprinkled throughout this patch for unknown,
incomplete, or might-need-optimization areas of the code.  Another
thing to note is that this does not introduce the logic needed for
following PHOT chains after single-page cleanup.

Signed-off-by: Greg Burd <greg@burd.me>