Eliminate the freelist from the buffer manager and depend on clock-sweep.
This set of changes removes the list of available buffers and instead
simply uses the clock-sweep algorithm to find and return an available
buffer.  While on the surface this appears to be removing an
optimization it is in fact eliminating code that induces overhead in the
form of synchronization that is problemmatic for multi-core systems.
This also changes the have_free_buffer() function to return true until
every buffer in the pool has been considered once by the clock-sweep
algorithm so as to inform the the pg_prewarm module as to when to stop
warming.