Eliminate the freelist from the buffer manager and depend on clocksweep.
This set of changes removes the list of available buffers and instead simply
uses the clocksweep algorithm to find and return an available buffer.  While on
the surface this appears to be removing an optimization it is in fact
eliminating code that induces overhead in the form of synchronization that is
problemmatic for multi-core systems.  This change removes the
have_free_buffer() function that was used in the pg_prewarm module.