Optimize modulo in clock-sweep algorithm
Improve performance by replacing the modulo operator which compiles into
a division instruction that can be slow on certain architectures.  When
the size of the clock (NBuffers) is a power-of-two we can simply
bitshift to get the modulo (4 instructions, ~3-4 cycles). When it isn't
we can replace modulo using a 64-bit multiplication by the inverse of
the clock size and a right shift as described in the paper "Division by
Invariant Integers using Multiplication" (4 instructions, ~8-12 cycles).
In both cases the branch prediction should be nearly 100% given that
NBuffers never changes at runtime.  In comparison a modulo operation
translates into IDIV and the code would require ~26-90 cycles. Switching
to this invariant method should use common ALU operations that don't
block the pipeline and have better instruction level parallelism.

[1] https://gmplib.org/~tege/divcnst-pldi94.pdf