Abstract and optimize clock-sweep buffer replacement algorithm
Re-author the clock-sweep algorithm such that it maintains its own state
and has a well defined API.  Improve performance by replacing the modulo
operator which compiles into a division instruction that can be slow
with two other approaches.  When the size of the clock (NBuffers) is a
power-of-two we can simply bitwise and to get the modulo (~1.42x speed
up).  When it isn't we can replace it using a 64-bit multiplication by
the inverse of the clock size and a right shift as described in the
paper "Division by Invariant Integers using Multiplication" (~1.11x
speed up). This second method is generated by GCC when the second
argument to a modulo is constant, however in our case it is not and so
we're left to implement this optimization ourselves.

[1] https://gmplib.org/~tege/divcnst-pldi94.pdf