Reorganize heap update logic and update simple_heap_update()
This commit refactors the interaction between heap_tuple_update(),
heap_update(), and simple_heap_update() to improve code organization and
flexibility. The changes are functionally equivalent to the previous
implementation except that now catalog tuples no longer depends on
HeapDetermineColumnsInfo() to determine which attributes changed, that
information is supplied to simple_heap_update().

As part of this reorganization, the handling of replica identity key
attributes has been adjusted. Instead of fetching a second copy of
the bitmap during an update operation, the caller is now required to
provide it. This change applies to both heap_update() and
heap_delete().